requirements.txt
discord.py>=2.4.0
python-dotenv==0.21.*
openai>=1.40.0
PyYAML>=6.1
dacite==1.6.*